backend/requirements.txt
fastapi[standard]
uvicorn
httpx
openai
groq
python-dotenv
newsapi-python